+++ scripts/test/dst.sh
@@ -11,3 +11,4 @@ export PATH="$(yarn bin):$PATH"
 env TZ=America/Sao_Paulo babel-node ./test/dst/parseISO/basic.js
 env TZ=Pacific/Apia babel-node ./test/dst/parseISO/samoa.js
 env TZ=Asia/Damascus babel-node ./test/dst/eachDayOfInterval/basic.js
+env TZ=America/Santiago babel-node ./test/dst/addBusinessDays/basic.js