+++ typings.d.ts
@@ -73,6 +73,9 @@ declare module 'date-fns' {
 // Regular Functions
 
 declare module 'date-fns' {
+  function addBusinessDays(date: Date | number, amount: number): Date
+  namespace addBusinessDays {}
+
   function addDays(date: Date | number, amount: number): Date
   namespace addDays {}
 
@@ -798,6 +801,11 @@ declare module 'date-fns' {
   const maxTime: number
 
   const minTime: number
+}
+
+declare module 'date-fns/addBusinessDays' {
+  import { addBusinessDays } from 'date-fns'
+  export default addBusinessDays
 }
 
 declare module 'date-fns/addDays' {
@@ -1580,6 +1588,11 @@ declare module 'date-fns/toDate' {
   export default toDate
 }
 
+declare module 'date-fns/addBusinessDays/index' {
+  import { addBusinessDays } from 'date-fns'
+  export default addBusinessDays
+}
+
 declare module 'date-fns/addDays/index' {
   import { addDays } from 'date-fns'
   export default addDays
@@ -2358,6 +2371,11 @@ declare module 'date-fns/subYears/index' {
 declare module 'date-fns/toDate/index' {
   import { toDate } from 'date-fns'
   export default toDate
+}
+
+declare module 'date-fns/addBusinessDays/index.js' {
+  import { addBusinessDays } from 'date-fns'
+  export default addBusinessDays
 }
 
 declare module 'date-fns/addDays/index.js' {
@@ -3143,6 +3161,9 @@ declare module 'date-fns/toDate/index.js' {
 // FP Functions
 
 declare module 'date-fns/fp' {
+  const addBusinessDays: CurriedFn2<number, Date | number, Date>
+  namespace addBusinessDays {}
+
   const addDays: CurriedFn2<number, Date | number, Date>
   namespace addDays {}
 
@@ -3761,6 +3782,11 @@ declare module 'date-fns/fp' {
   const maxTime: number
 
   const minTime: number
+}
+
+declare module 'date-fns/fp/addBusinessDays' {
+  import { addBusinessDays } from 'date-fns/fp'
+  export = addBusinessDays
 }
 
 declare module 'date-fns/fp/addDays' {
@@ -4663,6 +4689,11 @@ declare module 'date-fns/fp/toDate' {
   export = toDate
 }
 
+declare module 'date-fns/fp/addBusinessDays/index' {
+  import { addBusinessDays } from 'date-fns/fp'
+  export = addBusinessDays
+}
+
 declare module 'date-fns/fp/addDays/index' {
   import { addDays } from 'date-fns/fp'
   export = addDays
@@ -5561,6 +5592,11 @@ declare module 'date-fns/fp/subYears/index' {
 declare module 'date-fns/fp/toDate/index' {
   import { toDate } from 'date-fns/fp'
   export = toDate
+}
+
+declare module 'date-fns/fp/addBusinessDays/index.js' {
+  import { addBusinessDays } from 'date-fns/fp'
+  export = addBusinessDays
 }
 
 declare module 'date-fns/fp/addDays/index.js' {
@@ -6466,6 +6502,9 @@ declare module 'date-fns/fp/toDate/index.js' {
 // ECMAScript Module Functions
 
 declare module 'date-fns/esm' {
+  function addBusinessDays(date: Date | number, amount: number): Date
+  namespace addBusinessDays {}
+
   function addDays(date: Date | number, amount: number): Date
   namespace addDays {}
 
@@ -7191,6 +7230,11 @@ declare module 'date-fns/esm' {
   const maxTime: number
 
   const minTime: number
+}
+
+declare module 'date-fns/esm/addBusinessDays' {
+  import { addBusinessDays } from 'date-fns/esm'
+  export default addBusinessDays
 }
 
 declare module 'date-fns/esm/addDays' {
@@ -7973,6 +8017,11 @@ declare module 'date-fns/esm/toDate' {
   export default toDate
 }
 
+declare module 'date-fns/esm/addBusinessDays/index' {
+  import { addBusinessDays } from 'date-fns/esm'
+  export default addBusinessDays
+}
+
 declare module 'date-fns/esm/addDays/index' {
   import { addDays } from 'date-fns/esm'
   export default addDays
@@ -8751,6 +8800,11 @@ declare module 'date-fns/esm/subYears/index' {
 declare module 'date-fns/esm/toDate/index' {
   import { toDate } from 'date-fns/esm'
   export default toDate
+}
+
+declare module 'date-fns/esm/addBusinessDays/index.js' {
+  import { addBusinessDays } from 'date-fns/esm'
+  export default addBusinessDays
 }
 
 declare module 'date-fns/esm/addDays/index.js' {
@@ -9536,6 +9590,9 @@ declare module 'date-fns/esm/toDate/index.js' {
 // ECMAScript Module FP Functions
 
 declare module 'date-fns/esm/fp' {
+  const addBusinessDays: CurriedFn2<number, Date | number, Date>
+  namespace addBusinessDays {}
+
   const addDays: CurriedFn2<number, Date | number, Date>
   namespace addDays {}
 
@@ -10154,6 +10211,11 @@ declare module 'date-fns/esm/fp' {
   const maxTime: number
 
   const minTime: number
+}
+
+declare module 'date-fns/esm/fp/addBusinessDays' {
+  import { addBusinessDays } from 'date-fns/esm/fp'
+  export default addBusinessDays
 }
 
 declare module 'date-fns/esm/fp/addDays' {
@@ -11056,6 +11118,11 @@ declare module 'date-fns/esm/fp/toDate' {
   export default toDate
 }
 
+declare module 'date-fns/esm/fp/addBusinessDays/index' {
+  import { addBusinessDays } from 'date-fns/esm/fp'
+  export default addBusinessDays
+}
+
 declare module 'date-fns/esm/fp/addDays/index' {
   import { addDays } from 'date-fns/esm/fp'
   export default addDays
@@ -11954,6 +12021,11 @@ declare module 'date-fns/esm/fp/subYears/index' {
 declare module 'date-fns/esm/fp/toDate/index' {
   import { toDate } from 'date-fns/esm/fp'
   export default toDate
+}
+
+declare module 'date-fns/esm/fp/addBusinessDays/index.js' {
+  import { addBusinessDays } from 'date-fns/esm/fp'
+  export default addBusinessDays
 }
 
 declare module 'date-fns/esm/fp/addDays/index.js' {
@@ -14919,6 +14991,8 @@ declare module 'date-fns/esm/locale/zh-TW/index.js' {
 // dateFns Global Interface
 
 interface dateFns {
+  addBusinessDays(date: Date | number, amount: number): Date
+
   addDays(date: Date | number, amount: number): Date
 
   addHours(date: Date | number, amount: number): Date